+++ src/app.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import Page from './page';
+
+export default function App() {
+  return (
+    <Page />
+  );
+}
