+++ src/button.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function Button({ onClick, value }) {
+  return (
+    <button type="button" onClick={() => onClick(value)}>
+      {value}
+    </button>
+  );
+}
