+++ src/buttons.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import Button from './button';
+
+export default function Buttons({ onClick }) {
+  return (
+    <p>
+      {[1, 2, 3, 4, 5].map((i) => (
+        <Button
+          key={i}
+          value={i}
+          onClick={onClick}
+        />
+      ))}
+    </p>
+  );
+}
