+++ src/counter.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export default function Counter({ count, onClick }) {
+  return (
+    <button type="button" onClick={() => onClick(1)}>
+      Click me!
+      (
+      {count}
+      )
+    </button>
+  );
+}
